removed sleekxmpp Comm, renamed slixmpp to xmpp
