--- pyobs/comm/sleekxmpp/xep_0009/__init__.py
@@ -1,4 +1,0 @@
-from sleekxmpp.plugins.base import register_plugin
-from .rpc import XEP_0009
-
-register_plugin(XEP_0009)
