--- pyobs/comm/slixmpp/__init__.py
@@ -1,3 +1,0 @@
-from .xmppcomm import XmppComm
-
-__all__ = ["XmppComm"]
