--- pyobs/comm/slixmpp/xep_0009_timeout/stanza/__init__.py
@@ -1,9 +1,0 @@
-"""
-    SleekXMPP: The Sleek XMPP Library
-    Copyright (C) 2011 Nathanael C. Fritz, Dann Martens (TOMOTON).
-    This file is part of SleekXMPP.
-
-    See the file LICENSE for copying permission.
-"""
-
-from .RPC import MethodTimeout
