+++ pyobs/comm/xmpp/xmppcomm.py
